Add container tests, merge CI to single build_push.yaml (#230)
* feat: add container tests, merge CI to single build_push.yaml

Usage of the below features is described in the updated README.md

This was tested on a fork and is [working](https://github.com/ca-scribner/kubeflow-containers/actions/runs/636377739), but note that the fork did not have registry creds so it fails on the final step (that step will work once in regular repo).

Advice for review: Look at the `README.md` and follow the development pattern.  Check the `Makefile` and `build_push.yaml` for most changes.

Automated testing added for images.  These are broken into general (applies to all images) and specific (applies to a single image) tests.  These are driven by pytest and can be invoked through `make test/IMAGENAME`.

Testing follows the pattern of:
* locally run the given Docker image
* run tests in the running container

Structure and the tests themselves were modified from [docker-stacks](https://github.com/jupyter/docker-stacks/tree/master/test)

Most CI logic is now implemented in the `Makefile` and utilized by the GitHub Actions yaml.  This enables local testing that is nearly the same as the automated testing.

The `build.yaml`, `push.yaml`, and `push_on_comment.yaml` have been merged and refactored to reduce redundancy.  There is now a single `build_push.yaml` which:
* builds/scans/tests on all push to master or open/push to PR
* pushes images to ACR on all push to master or **if an open PR has the `auto-deploy` label**.  This is a simplification of the `build_on_comment.yaml` workflow, which gives controlled automated building within our CI (only team members can push to a repo and only team members can control labels, so both are tightly controlled)

The `Makefile` has been restructured to consolidate the CI and run the testing.  `make all` has been renamed to `make generate-dockerfiles` to make it more clear (`all` was not as intuitive once the `Makefile` was handling so much additional functionality.

Resolves StatCan/kubeflow-containers#228